Type fixes and prefetching for Visual Studio

+++ src/engine/eval.h
@@ -4,9 +4,9 @@
 #include "state.h"
 #include "hashtable.h"
 
-extern const int piecesquare[7][64];
+extern const short piecesquare[7][64];
 
-int EVAL_evaluate_board(const chess_state_t *s, hashtable_t *t);
-int EVAL_position_is_attacked(const chess_state_t *s, const int color, const int pos);
+short EVAL_evaluate_board(const chess_state_t *s, hashtable_t *t);
+int   EVAL_position_is_attacked(const chess_state_t *s, const int color, const int pos);
 
 #endif
